Refactor: split example plugins off to separate file
